Allow a custom index matcher
Some upstreams require some (arbitrary) headers
to be either indexed or not depending on when the
request is made. This is done so that application
developers can optimize the size of the dynamic
tables.

This PR allows the user of the library to specify
a custom index matching function which is called
for every header sent it, and therefore allows for
custom indexing or not based on requirements.